Revert "Hash TimetableVariants according to assigned IDs"
This reverts commit f3d936977a6ae800002041ebe3f3b72ddeecb05f.

Some variants as stored in the database are equal. My code relies on
hashing in sets and dictionaries to eliminate the duplicates, and
changing the hashing so that variants could be equal other than the IDs
and hash differently broke that duplicate elimination, greatly
increasing the number of calls the code made to section.contains(),
which is currently the bottleneck.