Only create one Point per BusSnapshot
Previously, the section.contains() method created a new point every time
it was called. Since it was being called many times per snapshot, this
resulted in many redundant Points being created, which according to
profiling was taking a long time. Now a Point is created once for each
snapshot and held onto.